[♻️Refactor]: wishList함수
