+++ src/lib/detail/generateInfoSection.js
@@ -1,0 +1,12 @@
+export function generateInfoSection(title, content, addContent = '') {
+  return content !== ''
+    ? /* html */ `
+      <div class="info_flex">
+        <dt class="min-w-32">${title}</dt>
+        <dd>
+          <p>${content}</p>
+          ${addContent}
+        </dd>
+      </div>`
+    : '';
+}
