+++ src/lib/detail/index.js
@@ -1,2 +1,4 @@
 export * from './insert.js';
 export * from './getNode.js';
+export * from './updateHeartIcon.js';
+export * from './generateInfoSection.js';
